Preserve executable bits found in source files
If we are making an archive from a source directory where there are files that are executable we
want to detect this automatically and preserve the executable bit in the archive without the
user having to intervene in any way.

Previous if the user wanted an entry to be executable they had to explicitly do this vs the
archiver just seeing the entry is executable and using that information.