Make file/dir permissions usable
In the case of some ZIP files the unix file mode is not available and if we unpack and repack a WAR file we need files to be readable and usable. In the case of unpackaging a Jenkins WAR file produced with the plexus-archiver the unix comes out as zero yielding and unreadable file. So it's possible I'm reading the information out of the zip entry incorrectly or it's not set correctly in the plexus-archiver. This needs to be tested.